First implementation
Read the iTunesLibrary. Put it into JSON with keys that are "nameFormat" for the iTunes XML keys "Name Format". Some keys the code in web_generator had previously found in iTunes are not available via this API (nor can I find them in my own iTunes XML file). Skip those.

Uses Swift Codable to encode the JSON, and it works quite nice. `Track` is in a separate file, and the main file has a category initializer that creates it from a `ITLibMediaItem`.

In my case this JSON is about 50% the size of the XML, but I don't have playlists in the JSON here.