Allow a directory to be passed to the program. (#10)
- If a directory is passed in as the first argument, a file with the name "iTunes-YYYY-MM-dd.json" will be created within it
- If no paramter is passed, it will continue to emit the json to stdout.

In Catalinea, this allows one to only allow this program instead of `sh` to have "Full Disk Access".

Previously I had a LaunchAgent Daemon that would back up iTunes on the first of every month. It broke with the update to Catalina, regardless of using this program. It would fire up a shell, make a string of the current date and have the program redirect into a backup file of the name with the date in it.

Now the Launch Agent will just invoke the program with a directory, and the shell is no longer necessary.