Add Destination.jsonGit
This will output a json file to a directory and add it to the git repository, if the directory is a git directory, and if the filename matches what is in the git directory.

This allows `git` to be able to write to the repository from `launchd`. This was the problem with https://github.com/bolsinga/bin_utils/issues/34