Improve handling duplicate runs of a git commit with the same base backup name.
- This way if the program runs a few times a day, it will handle each situation.
- If nothing has changed, the git tag with be "<name>-empty". If something has changed, but the basename is the same (since it is the same for each calendar day) it will be "<name>.01", "name.02" etc. Those names can be "<name>.01-empty" as well if that happens.

This is mostly because if the mac reboots and the LaunchAgent has the `RunAtLoad` key it will run at each reboot. In the old way, it will just re-write the same json.gz file over (but with any new data). This will not overwrite any data, and will not throw an error, causing the backup to fail.

This work and https://github.com/bolsinga/bin_utils/pull/43 should address https://github.com/bolsinga/bin_utils/issues/34