Add Patch/Patchable.trackCorrections
This is to be used when it is hard to get "the truth" from the current data. The first example is that Poster Childern "Grand Bargain! " album (note the '!' and ' ') and the single (first imported as "Grand Bargain!") are "similar" (since those two characters will be stripped). We need a "is equal" code path.

Additionally, this does not lend itself to a "the current data from iTunes is truth, figure it out" style of fix. This is because of the "isSimilar" code path.

Just hand-build TrackCorrection data. The TrackCorrection.Property enum may be added to to add more hand done corrections. This may be useful in the future.