Add IdentifierCorrection and Add Patch/Repair.replaceDurations (#683)
This is a much faster way; the batch query across the flat tracks
database has demonstrated that with very few exceptions, the
persistentID / itunesID remains consistent over the decades. This would
be a better way to do the old repairs, but this is what will be used
going forward. It may not be beneficial to change all that code now.