replaceDataAdded uses "first" dateAdded in history. (#706)
Looking at the data aggregated by the flat database, the dateAdded
property seems to be correct the first time it is added. It is only
after the TZ changes that it is entered incorrectly. Since this should
be an immutable value, use the first value added as the truth.

The general change for this Patch is that history, and not the current
library values, are the source of truth.