Backups can now be Patched
If an optional patch-url is given to the program, the URL should contain IdentifierCorrections. These corrections will be applied to the Tracks prior to backing up the data.

This way, errant data in the source iTunes data can be patched prior to backup. This is because iTunes has incorrect data within it that is non-editable (chaning dateAdded values is the first one covered).