Add replacePlay (#711)
This will look at the historical data. It will see if playDates are
moving forward. This seems to find many TZ off by an hour problems.

However it does not seem to be able to fix the "we have a playdate but
not a play count" problem. That will be addressed in a follow up.