Add CorrectionsDB (#744)
This database contains patches.
- It can be written to by the patch option
- It can be used by the backup option.
- It can be used by the repair option.

This database can be read by nightly backups and applied. This is
necessary since the Library may provide incorrect data, compared to the
historical truth that is currently backed up. It's useful to have in
this database, because with so many dates changing over history, they
all have to be corrected. The json files are too large, and the database
of corrections is smaller. It's also good going forward.

Fixes #727