Add ArchiveCommand. (#829)
- This will read flat databases from a git repository. It will iterate
through each and add it to a normalized archive database.
- Right now it only does artists. It can find what artists are new as
well as track changes to their name. The "later" versions of the name
will replace the older ones. This is just like the artist name repair.
- Adds artistids table, which is a lookup of itunesid -> artistid. This
will be used to create the albums and songs tables in future diffs. It
is possible this could be a TEMPORARY table; keeping it around for now.