Archive.plays
This will build up the archive with a lax play table schema! It handles the DST quirk by looking for dates that are exactly multiples of one hour apart. If they are exactly quirks, then it will use the last played date for the track. It will keep the newer count.