Limit quirks to less than a day difference (#842)
There were two songs in iTunes-V40-2006-12-01 where the playdate was an
exact multiple of the previous playdate, but one a day more than 24
hours away! This made it look like a quirk date (DST), so it would take
the old date, and the new count, and fail to add.

Fix is to check to see if the difference is less than 12 hours or not.